+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Support Monero Wallet RPC authentication.
+- Added `create-user` command.
+
 ### Changed
 
 - Added emoji count check to profile data validator.